Fix blend function to draw
Prior to this change, `gs_blend_function(GS_BLEND_ONE, GS_BLEND_ZERO)`
was called before rendering each source and the blend-function was not
set when blending the source. In addition, `gs_blend_state_push` was
called but `gs_blend_state_pop` was not called.
This commit fixes these bugs and use `gs_reset_blend_state` so that the
blend behavior is similar to the routine to render scene-items.
Close #7.